requirements.txt
fastapi>=0.111.0
uvicorn[standard]
sqlalchemy>=2.0
aiomysql
pymysql
alembic
python-dotenv
websockets
bson          # pip install pymongo 包含bson，或用独立bson包
httpx
pydantic-settings